cleanup comments and more refactor
