+++ compose_rl/ppo/callback.py
@@ -31,7 +31,7 @@ from transformers import PreTrainedTokenizer, PreTrainedTokenizerFast
 
 import compose_rl.utils as utils
 from compose_rl.ppo.buffer import MinibatchRolloutBuffer
-from compose_rl.ppo.generation_utils import hf_generate
+from compose_rl.ppo.generation_utils import hf_generate, vllm_generate
 from compose_rl.ppo.model import ComposerHFPolicyModel, ComposerMosaicPolicy
 from compose_rl.ppo.reward_manager import (
     ReferenceOutput,
@@ -60,191 +60,14 @@ __all__ = ['PPOCallback', 'env_reward']
 
 log = logging.getLogger(__name__)
 
-def vllm_generate(
-    # actor_critic: Policy,
-    vllm_engines: Optional[list],
-    # reward_manager: RewardManager,
-    batch: dict,
-    max_gen_len: int,
-    # precision: Precision,
-    # device_train_microbatch_size: int,
-    generation_kwargs: dict,
-    tokenizer: Tokenizer,
-    # eos_token_ids: list[int],
-) -> torch.Tensor:
-    """Run vllm generate on the prompts.
-
-    Runs generate over a set of sequences in the batch. It also does extra computation
-    that is required for later loss computation.
-
-    Args:
-        actor_critic (ComposerMosaicPolicy): Actor critic model to run generate over.
-        reward_manager (RewardManager): Composes the reference IFT model and all generate models.
-        batch (dict): The batch of data to run generate over.
-        max_gen_len (int): Maximum generation length.
-        precision (Precision): Precision to run computation.
-        device_train_microbatch_size (int): Device train microbatch size for the training job.
-            We need to do all log_prob computation with this in order to maintain numerics.
-        generation_kwargs (dict): Generation keyword arguments.
-        tokenizer (Tokenizer): The actor critic's tokenizer.
-        eos_token_ids (list[int]): A list of eos token ids.
-
-    Returns:
-        sequences (tensor): Tensor containing the prompt and generated sequences.
-            The shape of the tensor is [batch_size, prompt_len + max_gen_len].
-
-    Note:
-        Use the .get() method on an AsyncResult object (see Returns, above) to resolve it.
-    """
-    # 1. Gather all prompts from all ranks
-    # 2. Run generate over all prompts in one go
-    # 3. Scatter the generated responses back to the correct rank
-    # 4. Save the tokenized sequences in the batch for future use
-    pad_token_id = tokenizer.pad_token_id  # type: ignore
-    cur_device = batch['prompt'].device
-    prompt_tokens = batch['prompt']
-    # Pull the necessary variables from the batch and self
-    prompt_all_gather_start_time = time.time()
-
-    all_batched_prompts = dist.all_gather_object(prompt_tokens)
-    batch_sizes = [len(batch) for batch in all_batched_prompts]
-
-    log.info(
-        f'took : {time.time() - prompt_all_gather_start_time} to gather prompts',
-    )
-    all_prompts = [
-        prompt for batch in all_batched_prompts for prompt in batch
-    ]
-
-    if dist.get_global_rank() == 0:
-        futs = []
-        sampling_params = {
-            'temperature': generation_kwargs.get('temperature', 1.0),
-            'top_p': generation_kwargs.get('top_p', 1.0),
-            'top_k': generation_kwargs.get('top_k', 50),
-            'max_tokens': max_gen_len,
-        }
-
-        # We have to remove all pad tokens here
-        all_prompts = [[
-            token
-            for token in prompt.detach().cpu().tolist()
-            if token != pad_token_id
-        ]
-                    for prompt in all_prompts]
-
-        # Generate with vllm
-        # Calculate the base batch size
-        vllm_batch_size = len(all_prompts) // len(vllm_engines)
-        # Calculate the remainder (prompts that don't fit evenly)
-        remainder = len(all_prompts) % len(vllm_engines)
-
-        start_idx = 0
-        for i, engine in enumerate(vllm_engines):
-            # Assign one extra prompt to the first 'remainder' engines
-            if i < remainder:
-                end_idx = start_idx + vllm_batch_size + 1
-            else:
-                end_idx = start_idx + vllm_batch_size
-
-            cur_prompts_ids = all_prompts[start_idx:end_idx]
-            cur_prompts = [
-                tokenizer.decode(prompt)  # type: ignore
-                for prompt in cur_prompts_ids
-            ]
-
-            futs.append(
-                engine.generate.remote(
-                    cur_prompts,
-                    sampling_params=sampling_params,
-                ),
-            )
-
-            # Update the start index for the next iteration
-            start_idx = end_idx
-
-        start_time = time.time()
-        results = ray.get(futs)
-        all_responses = []
-
-        # Get all of the ray futures
-        for i, result in enumerate(results):
-            # Each result is a list of responses this assumes one output per input
-            all_responses.extend([
-                resp.outputs[0].token_ids for resp in result
-            ])
-
-        log.info(
-            f'took: {time.time() - start_time} to gather futures',
-        )
-
-        # Distribute padded responses back to the correct device
-        split_responses = []
-        start = 0
-        for size in batch_sizes:
-            split_responses.append(
-                all_responses[start:start + size],
-            )
-            start += size
-    else:
-        # Remove the memory from all gather as they are only used for the first rank
-        all_batched_prompts = None
-        all_prompts = None
-        split_responses = None
-
-    # Do another garbage collection and empty the cache
-    gc.collect()
-    if torch.cuda.is_available():
-        torch.cuda.empty_cache()
-
-    dist.barrier()
-
-    # Scatter the generated responses back to the correct rank
-    local_responses = [None]
-    start_time = time.time()
-    torch.distributed.scatter_object_list(
-        local_responses,
-        split_responses,
-        src=0,
-    )
-    local_responses = local_responses[0]
-
-    log.info(f'took: {time.time() - start_time} to scatter prompts')
-
-    # Pad the responses to the maximum length
-    max_vllm_generated_len = max([
-        len(response) for response in local_responses  # type: ignore
-    ])
-    padded_responses = []
-    for sequence in local_responses:  # type: ignore
-        sequence = list(sequence)
-        if len(sequence) < max_vllm_generated_len:
-            sequence = sequence + [
-                pad_token_id,
-            ] * (max_vllm_generated_len - len(sequence))
-
-        padded_responses.append(sequence)
-
-    # Convert the padded responses to a tensor
-    padded_responses = torch.tensor(
-        padded_responses,
-        dtype=prompt_tokens.dtype,
-        device=cur_device,
-    )
-
-    # Construct full sequences from the prompt and padded responses
-    sequences = torch.cat([prompt_tokens, padded_responses], dim=-1)
-    return sequences
 
 def env_reward(
     actor_critic: Policy,
-    vllm_engines: Optional[list],
     reward_manager: RewardManager,
     batch: dict,
     max_gen_len: int,
     precision: Precision,
     device_train_microbatch_size: int,
-    generation_kwargs: dict,
     tokenizer: Tokenizer,
     eos_token_ids: list[int],
     kl_estimator: str,
@@ -267,7 +90,6 @@ def env_reward(
         precision (Precision): Precision to run computation.
         device_train_microbatch_size (int): Device train microbatch size for the training job.
             We need to do all log_prob computation with this in order to maintain numerics.
-        generation_kwargs (dict): Generation keyword arguments.
         tokenizer (Tokenizer): The actor critic's tokenizer.
         eos_token_ids (list[int]): A list of eos token ids.
         kl_estimator (str): Which kl estimator to use. Options are 'k1', 'k2', 'k3' and 'k3_offpolicy'.
@@ -820,16 +642,11 @@ class PPOCallback(CallbackWithConfig):
             # If vllm engines are available, we use them to generate sequences in one go
             if self.vllm_engines is not None:
                 sequences = vllm_generate(
-                    # actor_critic=self.actor_critic,
                     vllm_engines=self.vllm_engines,
-                    # reward_manager=self.reward_manager,
                     batch=batch,
                     max_gen_len=max_gen_len,
-                    # precision=self.precision,
-                    # device_train_microbatch_size=self.device_train_microbatch_size,
                     generation_kwargs=generation_kwargs,
                     tokenizer=self.tokenizer,  # type: ignore
-                    # eos_token_ids=self.eos_token_ids,  # type: ignore
                 )
             else:
                 # Go the HF policy generate route
@@ -846,13 +663,15 @@ class PPOCallback(CallbackWithConfig):
                         idx=i,
                         minibatch_size=self.device_generate_batch_size,
                     )
+
                     gen_sequences = hf_generate(
-                        self.actor_critic,
-                        max_gen_len,
-                        gen_batch,
-                        pad_token_id, # type: ignore
-                        generation_kwargs,
+                        actor_critic=self.actor_critic,
+                        max_gen_len=max_gen_len,
+                        batch=gen_batch,
+                        pad_token_id=pad_token_id, # type: ignore
+                        generation_kwargs=generation_kwargs,
                     )
+
                     all_sequences.append(gen_sequences)
                 # Add right padding to all sequences and concatenate them
                 max_len = max([seq.shape[1] for seq in all_sequences])
@@ -870,7 +689,6 @@ class PPOCallback(CallbackWithConfig):
 
         env_outputs, prompts_and_gens, ref_outputs, all_rewards_dict = env_reward(
             actor_critic=self.actor_critic,  # pyright: ignore
-            vllm_engines=self.vllm_engines,
             reward_manager=self.reward_manager,
             batch=batch,
             max_gen_len=self.max_gen_len,
@@ -952,7 +770,6 @@ class PPOCallback(CallbackWithConfig):
             output_minibatch (dict): The final minibatch from the environment, with all AsyncResult
                 objects resolved and outputs processed for PPO training.
         """
-        outputs = []
         env_outs, ref_outs, rew_dict = partial_outputs
         rew_outs = self.reward_manager.resolve_outputs(
             ref_output=ref_outs,
@@ -962,23 +779,12 @@ class PPOCallback(CallbackWithConfig):
             center_reward_mean=self.center_reward_mean,
         )
         env_outs.update(rew_outs)
-        outputs.append(env_outs)
 
-        env_outputs = {}
-        # Repadding all observations, right padded attention masks to the same length for composer.
-        max_len = max([output['obs'].shape[-1] for output in outputs])
-        for output in outputs:
-            output['obs'] = add_right_padding(
-                output['obs'],
-                max_len,
-                self.pad_token_idx,  # type: ignore
-            )
-            output['right_padded_attn_mask'] = torch.logical_not(
-                torch.eq(output['obs'], self.pad_token_idx),  # type: ignore
-            )
-
-        for key in outputs[0].keys():
-            env_outputs[key] = torch.cat([output[key] for output in outputs])
+        # Adding the right_padded_attn_mask to the env_outputs
+        env_outs['right_padded_attn_mask'] = torch.logical_not(
+            torch.eq(env_outs['obs'], self.pad_token_idx),  # type: ignore
+        )
+        env_outputs = env_outs
 
         # Now that rewards are resolved, we can compute advantages
         env_outputs['advantages'] = compute_advantages(
